pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "followin-mcp"
version = "0.1.0"
description = "A crypto-focused Followin MCP prototype with normalization, clustering, ranking, and debug tooling."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
followin-mcp-dev = "followin_mcp.dev:main"
followin-mcp-server = "followin_mcp.mcp.server:main"
followin-mcp-web = "followin_mcp.demo.webapp:main"

[tool.setuptools]
packages = [
    "followin_mcp",
    "followin_mcp.core",
    "followin_mcp.demo",
    "followin_mcp.mcp",
]
